Inlining checksums along side the download links
The checksum file has an extension of .checksum, which is very inconvenient to open in a GUI editor. Also, since it is a java properties file, it is not very user friendly. Many sites provide the checksums along side the downloads and being the general pattern, I thought we could move to that. The 'Checksum' header links to the downloadable file.